+++ src/Command/GenerateCommand.php
@@ -46,7 +46,7 @@ class GenerateCommand extends FormattedOutputCommand
 			->addArgument('env', InputArgument::REQUIRED, 'Target environment code')
 			->addArgument('dbchange', InputArgument::OPTIONAL, 'DbChange code of the dbChange to be generated')
 			->addArgument('fragmentIndex', InputArgument::OPTIONAL, 'Fragment index of the fragment to be generated')
-			->addOption('debug', 'd', InputOption::VALUE_NONE, 'y');
+			->addOption('debug', 'd', InputOption::VALUE_NONE, '');
 	}
 
 	protected function execute(InputInterface $input, OutputInterface $output)